feat : bid register 구현중
