--- com.nameslowly.coinauctions.bidwin/src/main/java/com/nameslowly/coinauctions/bidwin/application/service/UserService.java
@@ -1,8 +1,0 @@
-package com.nameslowly.coinauctions.bidwin.application.service;
-
-import com.nameslowly.coinauctions.bidwin.application.dto.UserDto;
-
-public interface UserService {
-
-    UserDto getUser();
-}
