--- com.nameslowly.coinauctions.bidwin/src/main/java/com/nameslowly/coinauctions/bidwin/infrastructure/AuctionServiceImpl.java
@@ -1,15 +1,0 @@
-package com.nameslowly.coinauctions.bidwin.infrastructure;
-
-import com.nameslowly.coinauctions.bidwin.infrastructure.auction.AuctionDto;
-import com.nameslowly.coinauctions.bidwin.infrastructure.auction.AuctionService;
-import org.springframework.stereotype.Component;
-
-@Component
-public class AuctionServiceImpl implements AuctionService {
-
-    @Override
-    public AuctionDto getAuction() {
-
-        return null;
-    }
-}
