+++ com.nameslowly.coinauctions.bidwin/src/main/java/com/nameslowly/coinauctions/bidwin/infrastructure/auction/AuctionFeignClient.java
@@ -2,10 +2,12 @@ package com.nameslowly.coinauctions.bidwin.infrastructure.auction;
 
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.context.annotation.Primary;
+import org.springframework.web.bind.annotation.GetMapping;
 
 @FeignClient(name = "auction-service", fallbackFactory = AuctionFallbackFactory.class)
 @Primary
 public interface AuctionFeignClient extends AuctionService {
 
-
+    @GetMapping("/aoi/auctions")
+    AuctionDto getAuction();
 }
