+++ com.nameslowly.coinauctions.bidwin/src/main/java/com/nameslowly/coinauctions/bidwin/infrastructure/auction/AuctionService.java
@@ -2,6 +2,5 @@ package com.nameslowly.coinauctions.bidwin.infrastructure.auction;
 
 public interface AuctionService {
 
-
     AuctionDto getAuction();
 }
