+++ com.nameslowly.coinauctions.bidwin/src/main/java/com/nameslowly/coinauctions/bidwin/infrastructure/coinpay/CoinpayFeignClient.java
@@ -2,10 +2,12 @@ package com.nameslowly.coinauctions.bidwin.infrastructure.coinpay;
 
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.context.annotation.Primary;
+import org.springframework.web.bind.annotation.GetMapping;
 
 @FeignClient(name = "coinpay-service", fallbackFactory = CoinpayFallbackFactory.class)
 @Primary
 public interface CoinpayFeignClient extends CoinpayService {
 
-
+    @GetMapping("/api/coinpays")
+    CoinDto getCoin();
 }
