+++ com.nameslowly.coinauctions.bidwin/src/main/java/com/nameslowly/coinauctions/bidwin/infrastructure/coinpay/CoinpayService.java
@@ -2,6 +2,5 @@ package com.nameslowly.coinauctions.bidwin.infrastructure.coinpay;
 
 public interface CoinpayService {
 
-
     CoinDto getCoin();
 }
