+++ com.nameslowly.coinauctions.bidwin/src/main/java/com/nameslowly/coinauctions/bidwin/infrastructure/user/UserDto.java
@@ -1,4 +1,4 @@
-package com.nameslowly.coinauctions.bidwin.infrastructure.auction;
+package com.nameslowly.coinauctions.bidwin.infrastructure.user;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,8 +7,7 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class AuctionDto {
+public class UserDto {
 
-    private Long auctionId;
-
+    private Long userId;
 }
