+++ com.nameslowly.coinauctions.bidwin/src/main/java/com/nameslowly/coinauctions/bidwin/infrastructure/user/UserFallback.java
@@ -1,15 +1,15 @@
-package com.nameslowly.coinauctions.bidwin.infrastructure.coinpay;
+package com.nameslowly.coinauctions.bidwin.infrastructure.user;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 
 @Slf4j
 @Component
-public class CoinpayFallback implements CoinpayFeignClient {
+public class UserFallback implements UserFeignClient {
 
     @Override
-    public CoinDto getCoin() {
-        log.info("coin service error");
+    public UserDto getUser() {
+        log.info("user service error");
 //        throw new RuntimeException("user service error");
         return null;
     }
