+++ com.nameslowly.coinauctions.bidwin/src/main/java/com/nameslowly/coinauctions/bidwin/infrastructure/user/UserFallbackFactory.java
@@ -1,0 +1,13 @@
+package com.nameslowly.coinauctions.bidwin.infrastructure.user;
+
+import org.springframework.cloud.openfeign.FallbackFactory;
+import org.springframework.stereotype.Component;
+
+@Component
+public class UserFallbackFactory implements FallbackFactory<UserFeignClient> {
+
+    @Override
+    public UserFeignClient create(Throwable cause) {
+        return new UserFallback();
+    }
+}
