+++ com.nameslowly.coinauctions.bidwin/src/main/java/com/nameslowly/coinauctions/bidwin/infrastructure/user/UserService.java
@@ -1,0 +1,6 @@
+package com.nameslowly.coinauctions.bidwin.infrastructure.user;
+
+public interface UserService {
+
+    UserDto getUser();
+}
