+++ com.nameslowly.coinauctions.bidwin/src/main/resources/application-dev.yml
@@ -23,6 +23,27 @@ spring:
     username: guest
     password: guest
 
+  cloud:
+    openfeign:
+      circuit breaker:
+        enabled: true
+        alphanumeric-ids:
+          enabled: true
+
+  resilience4j:
+    circuit breaker:
+      configs:
+        default:
+          registerHealthIndicator: true
+          slidingWindowType: COUNT_BASED
+          slidingWindowSize: 5
+          minimumNumberOfCalls: 5
+          slowCallRateThreshold: 100
+          slowCallDurationThreshold: 60000
+          failureRateThreshold: 50
+          permittedNumberOfCallsInHalfOpenState: 3
+          waitDurationInOpenState: 20s
+
 message:
   exchange: app
   queue:
